+++ plugin_ide.ui/src/melnorme/lang/ide/ui/ContentAssistConstants.java
@@ -11,13 +11,7 @@
 package melnorme.lang.ide.ui;
 
 import melnorme.lang.ide.ui.text.completion.ContentAssistConstants_Default;
-import _org.eclipse.dltk.ui.PreferenceConstants;
 
 public interface ContentAssistConstants extends ContentAssistConstants_Default {
-	
-	String PROPOSALS_FOREGROUND = PreferenceConstants.CODEASSIST_PROPOSALS_FOREGROUND;
-	String PROPOSALS_BACKGROUND = PreferenceConstants.CODEASSIST_PROPOSALS_BACKGROUND;
-	String PARAMETERS_FOREGROUND = PreferenceConstants.CODEASSIST_PARAMETERS_FOREGROUND;
-	String PARAMETERS_BACKGROUND = PreferenceConstants.CODEASSIST_PARAMETERS_BACKGROUND;
 	
 }