+++ plugin_ide.ui/src/mmrnmhrm/ui/preferences/DeeEditorConfigurationBlock.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007, 2014 IBM Corporation and others.
+ * Copyright (c) 2015 Bruno Medeiros and other Contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -10,55 +10,13 @@
  *******************************************************************************/
 package mmrnmhrm.ui.preferences;
 
-import melnorme.lang.ide.ui.preferences.EditorAppearanceColorsComponent.EditorColorItem;
 import melnorme.lang.ide.ui.preferences.LangEditorConfigurationBlock;
-import melnorme.lang.ide.ui.preferences.PreferencesMessages;
 import melnorme.lang.ide.ui.preferences.common.AbstractComponentsPrefPage;
-import melnorme.util.swt.components.fields.CheckBoxField;
-
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.widgets.Composite;
-
-import _org.eclipse.dltk.ui.PreferenceConstants;
 
 public class DeeEditorConfigurationBlock extends LangEditorConfigurationBlock {
 	
 	public DeeEditorConfigurationBlock(AbstractComponentsPrefPage mainPreferencePage) {
 		super(mainPreferencePage);
-	}
-	
-	@Override
-	protected void createAppearanceGroup(Composite parent) {
-		createBooleanField(parent,
-			PreferenceConstants.EDITOR_MATCHING_BRACKETS,
-			new CheckBoxField(PreferencesMessages.EditorPreferencePage_matchingBrackets));
-		
-		super.createAppearanceGroup(parent);
-	}
-	
-	@Override
-	protected EditorColorItem[] createEditorAppearanceColorEntries() {
-		return new EditorColorItem[] {
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_matchingBracketsHighlightColor,
-					PreferenceConstants.EDITOR_MATCHING_BRACKETS_COLOR),
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_backgroundForMethodParameters,
-					PreferenceConstants.CODEASSIST_PARAMETERS_BACKGROUND),
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_foregroundForMethodParameters,
-					PreferenceConstants.CODEASSIST_PARAMETERS_FOREGROUND),
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_backgroundForCompletionReplacement,
-					PreferenceConstants.CODEASSIST_REPLACEMENT_BACKGROUND),
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_foregroundForCompletionReplacement,
-					PreferenceConstants.CODEASSIST_REPLACEMENT_FOREGROUND),
-				new EditorColorItem(
-					PreferencesMessages.EditorPreferencePage_sourceHoverBackgroundColor,
-					PreferenceConstants.EDITOR_SOURCE_HOVER_BACKGROUND_COLOR,
-					PreferenceConstants.EDITOR_SOURCE_HOVER_BACKGROUND_COLOR_SYSTEM_DEFAULT,
-					SWT.COLOR_INFO_BACKGROUND) };
 	}
 	
 }