+++ documentation/ChangeLog.md
@@ -1,6 +1,9 @@
 ## DDT release ChangeLog
 
 ### (NextVersion)
+ * Added support for Eclipse dark theme. Namely:  
+   * Syntax/source coloring now have different settings for dark theme.
+   * Fixed tools console colors, as well as content assist pop-up colors.
  * Fixed: `foreach(const ref x; y)` is invalid syntax. (#117)
  * Fixed: Unindent (Shift-Tab) broken, does nothing after empty lines in selection. (#116)